requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.0.3
streamlit>=1.28.0
PyPDF2>=3.0.1
google-generativeai>=0.3.0